rename: 사이드바 최하위 컴포넌트 구조를 '컴포넌트명/index.tsx'에서 '컴포넌트명.tsx' 로 수정 (#104)
