+++ src/components/Layout/SideBar/BrandLogo.tsx
@@ -1,8 +1,14 @@
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import { URL } from '../../../../constant';
-import { colors, padding, border } from '../../../../styles';
-import PokemonLogo from '../../../../assets/pokemonBall.svg';
+import { URL } from '../../../constant';
+import { colors, padding } from '../../../styles';
+import LogoImage from '../../../assets/pokemonBall.svg';
+
+interface BrandLogoProps {
+  style: {
+    padding: string;
+  };
+}
 
 const BrandLogo = ({ style }: BrandLogoProps) => {
   const { padding } = style;
@@ -11,7 +17,7 @@ const BrandLogo = ({ style }: BrandLogoProps) => {
     <Link to={URL.index.link}>
       <S.Logo padding={padding}>
         <div className="logo-image">
-          <img src={PokemonLogo} alt="Pokemon ERP" />
+          <img src={LogoImage} alt="" />
         </div>
         <div className="logo-text">Pokemon ERP</div>
       </S.Logo>
@@ -19,22 +25,12 @@ const BrandLogo = ({ style }: BrandLogoProps) => {
   );
 };
 
-type BrandLogoProps = {
-  style: {
-    padding: string;
-  };
-};
-
-type LogoProps = {
-  padding: string;
-};
-
 const S = {
-  Logo: styled.article<LogoProps>`
+  Logo: styled.article<BrandLogoProps['style']>`
     display: flex;
     align-items: center;
+    /* justify-content: center; */
     gap: ${padding.md};
-    border-bottom: ${border.default};
     height: 80px;
     padding: ${(props) => props.padding};
 
@@ -47,9 +43,10 @@ const S = {
     }
 
     .logo-text {
-      color: ${colors.semantic.primary};
+      color: ${colors.semantic.text.dark};
       font-weight: 500;
-      font-size: 24px;
+      font-size: 20px;
+      line-height: 1;
     }
   `,
 };
